requirements.txt
pytest==8.2.2
pytest-cov==5.0.0
requests==2.32.3
selenium==4.23.1
coverage==7.6.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker~=26.3.0